Extend integration test to play web audio
